move workspace context files
